+++ eva/src/js/componentsSettings.js
@@ -318,6 +318,7 @@ export default {
       'defaultSourceDataUpdates',
       'resetValuesWhichAreNot',
       'searchMode',
+      'selectedValuesAbove',
     ],
     picker: [
       'showLastTimeBlock',
@@ -996,6 +997,12 @@ export default {
     {
       option: 'resetValuesWhichAreNot',
       description: 'Сбросить значения, если источник данных их не содержит',
+      elem: 'switch',
+      default: false,
+    },
+    {
+      option: 'selectedValuesAbove',
+      description: 'Выбранные значения закрепить сверху',
       elem: 'switch',
       default: false,
     },